Move GeneratedJaxRsResourceGizmoAdaptor to SPI
